libft

+++ ft_bzero.c
@@ -6,11 +6,13 @@
 /*   By: hboutale <hboutale@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/30 13:05:56 by hboutale          #+#    #+#             */
-/*   Updated: 2024/10/30 13:45:02 by hboutale         ###   ########.fr       */
+/*   Updated: 2024/10/31 14:06:27 by hboutale         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_bzero(void *ptr, int size)
+#include "libft.h"
+
+void	ft_bzero(void *ptr, size_t size)
 {
 	int i;
 	unsigned char *str;
