+++ ft_memcpy.c
@@ -6,11 +6,13 @@
 /*   By: hboutale <hboutale@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/30 13:20:42 by hboutale          #+#    #+#             */
-/*   Updated: 2024/10/30 13:54:31 by hboutale         ###   ########.fr       */
+/*   Updated: 2024/10/31 14:46:52 by hboutale         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memcpy(void *dest, void *src, int size)
+#include "libft.h"
+
+void	*ft_memcpy(void *dest, void *src, size_t size)
 {
 	int i;
 	unsigned char *dest_ptr = dest;
