+++ ft_memset.c
@@ -6,11 +6,13 @@
 /*   By: hboutale <hboutale@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/30 12:07:18 by hboutale          #+#    #+#             */
-/*   Updated: 2024/10/30 13:46:41 by hboutale         ###   ########.fr       */
+/*   Updated: 2024/10/31 14:48:21 by hboutale         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *ptr, int ch, int n)
+#include "libft.h"
+
+void	*ft_memset(void *ptr, int ch, size_t n)
 {
 	int i;
 	unsigned char c;
