+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: hboutale <hboutale@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/30 11:53:23 by hboutale          #+#    #+#             */
-/*   Updated: 2024/10/30 11:55:57 by hboutale         ###   ########.fr       */
+/*   Updated: 2024/10/31 15:03:27 by hboutale         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(char *s)
+#include "libft.h"
+
+size_t	ft_strlen(char *s)
 {
-	int count;
+	size_t count;
 
 	count = 0;
 	while (*s != '\0')
